Добавлен файл render-comment.js, исправлены мелкие ошибки.
